# Optimize wasm.(*ModuleInstance).BuildFunctions