Before:
```sh
❯ go test -benchmem -memprofile memprofile.out -cpuprofile profile.out -run=^$ -bench ^BenchmarkInitialization/compiler$ github.com/tetratelabs/wazero/internal/integration_test/bench
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/compiler-10                69037             18685 ns/op          145951 B/op        128 allocs/op
PASS
ok      github.com/tetratelabs/wazero/internal/integration_test/bench   2.850s
```
<img width="1341" alt="Screenshot 2022-11-05 at 13 36 14" src="https://user-images.githubusercontent.com/1851985/200140281-90167063-06eb-41f9-b0e4-c28ebb0e12d7.png">


After:
```sh
❯ go test -benchmem -memprofile memprofile.out -cpuprofile profile.out -run=^$ -bench ^BenchmarkInitialization/compiler$ github.com/tetratelabs/wazero/internal/integration_test/bench
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
BenchmarkInitialization/compiler-10                64510             17417 ns/op          145950 B/op        128 allocs/op
PASS
ok      github.com/tetratelabs/wazero/internal/integration_test/bench   1.600s
```
<img width="1335" alt="Screenshot 2022-11-05 at 13 39 00" src="https://user-images.githubusercontent.com/1851985/200140452-57ea1161-945d-4181-9f54-b7e31ef49db9.png">

Not a massive win but simpler code IMO